requirements.txt
django~=4.0
djoser~=2.1.0
django-rest-framework
django-cors-headers
psycopg2-binary

djangorestframework~=3.13.1

gunicorn
asgiref
pytz
Pyrebase~=3.0.27

six~=1.16.0